% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockSizeCalibrate}
\alias{blockSizeCalibrate}
\title{calculate the optimal block size}
\usage{
blockSizeCalibrate(ret, b_vec = as.integer(c(1, 3, 6, 10)), alpha = 0.05,
  M = 199L, K = 1000L, b_av = 5L, T_start = 50L)
}
\arguments{
\item{ret}{refers to a returns matrix (columns = 2)}

\item{b_vec}{see vignette}

\item{alpha}{see vignette}

\item{M}{see vignette}

\item{K}{see vignette}

\item{b_av}{see vignette}

\item{T_start}{see vignette}
}
\value{
a list with probabilities and optimal block size
}
\description{
by Algorithm 3.1 in vignette on page 5
}

